sort_size4de

+++ src/sort_size5.c
@@ -1,7 +1,8 @@
 #include "../include/push_swap.h"
+
 void	sort_size5(t_list **stack_a, t_list **stack_b)
 {
-	int	distance;
+	size_t	distance;
 
 	if (is_sorted(stack_a))
 		return ;
